Introducing: the engine concept
Sider has been tied to redis, but the time
has come to expand the concept onto greater
things. Like other databases.

Sider is now aware that there are different
databases that can be loaded and run. I e I had
the need to add postgres to this.

NB
This is a breaking change - you need to
move any existing ~/.sider/images and
~/.sider/containers to snapshots
and dbs respectively. Or create
a .siderrc and set snapshotsFolder
to images and dbsFolder to containers
respectively.

Test plan:
* sider base command still works.
* sider works without any ~/.sider
  folder.
* Adding a snapshot with the redis-engine
  works. Verify that the size of the snapshot corresponds
  to the imported file.
* sider snapshot list lists all snapshots
  with their respective engine.
* sider snapshot remove works.
* sider db start to clone an image with and without port works.
* sider db start with exsting db works.
* sider db start with another port works.
* sider db reset works.
* sider db promote works.
* sider db list lists all dbs with their
  images and associated engines.