Use settings to store ports in dbs
Ports are really just another type of setting
(although it will always be defined).

Move settings so it can be stored per engine,db and
overridden on the cli as a run once parameter.

Rename defaultPort to only port.

NB
Will add persisting of settings on
a db in a follwing commit.
There is no longer a port in the sider db list.
Will be added later.
Update README.md

Test plan:
* Hijack file-db and run the cloneSnapshotToDb,
  getEngineConfig, setEngineConfig, getDbConfig
  and setDbConfig in isolation. Verify db formats
  are still listed correctly after port removal and
  cloning works
* Use sider-engine to set configs and get configs
  on some engine.
* Hijack engines/index and start an engine with
  cliConfig and no cliConfig. Verify port
  is honored if set in cliConfig and that
  it trumps db port which in turn trumps
  the default port on the engine.
* Clone a postgres snapshot a port=6380 parameter on
  the command line when starting the db. Verify
  that parameter is honored and db is started on that
  port.
* Start the previously cloned db above. Should revert
  to default port.
* Start the previously cloned db above again, this
  time with a runtime-parameter (port=1235) verify
  runtime parameter is honored.
* Do the same with a redis this time also adding
  version=3.2.6 and omitting it, making sure
  redis version is honored.