Refactor engine to be consistent with db config cmds
Refactor and set engine config as it
will be in sider db setconf

Frowning a bit on non natural language
set-config is too much to write, so opting
for setconf as it's shorter, but noting
the frown and grumpy feeling of doing so.

Rename set get and remove on engine to
setconf getconf and remconf so we're
consistent.

Test plan:
* Run sider engine setconf, getconf and remconf
  and verify no regression (e g values are
  set getted and removed as before).