Adding tab-tab completion to sider
Commander does not provide command-line
completion as of yet. Make some
format that can be somewhat shared
between completion and commander to
avoid duplication.

Also: I want the completion to be as fast
as possible. There's an annoying short
but still noticeable delay
when using tab-tab (or anything not
being an in-memory bash function I
assume).

Therefore to micro-optimize some: use
inline require:s in sider.js
because it's faster (timed it with an
hrtimer).

Test plan:
* Start with a blank shell, type sider install-completion.
  Follow the instructions, exit the shell and verify
  sider db start can now be completed.
* Uninstall the completion by typing sider uninstall-completion,
  exit the shell and verify that sider now does not complete.
* Type verify options are added after sider db start completion,
  that existing dbs are listed when tabbed and that snapshots are
  listed with and without command line parameters (-p, --persist).