Replace cli-table with my own
cli-table was nice but
for keeping the package
count low + I want it
to shrink when terminal size
is less than available space.

The first row is treated as a
heading - the rest are data-rows.

Table will shrink to min 5
(or less if the column is less in width)
but not further. Text is ellipsised
if is needs to shrink to accomodate
to the terminal width.

All is broken out if I would - y'know
want do break this out to it's own module.

Test plan:
* One cell displays ok.
* One cell and one row displays ok.
* Cells above 5 chars are shrunk proportionally-ish
  to the their respective length
  when (do this by scaling down the terminal
  one char at a time and printing the table).
  Scaling stops if the columns are equal to
  or less than 5. Do this for tables with 1, 2
  and many columns.