sider db list
This time dates that actually work.

The table now accepts a function
that gets access to the entire dataset.

This way headers can be set to terminal colors
(if supported) as long as it doesn't manipulate
the visible length of characters (cannot
set terminal control characters before
calculating width as they take up bytes
but not visible width on the screen).

Test plan:
* Usual help tests, except there is
  no positional arguments and thusly
  help is not printed when executing
  the empty command.
* No databases displays empty headings.
  Verify headings only are visible.
* Add a new database but do not start
  it. Verify name, snapshot name, engine
  and crated displays correct times.
  Verify last used says never.
* Add one more database and start it.
  List databases and verify start and
  last usage time correct. Start it again,
  list again and verify last used updates.