Fix nomenclature: runtimeConfig and it's spec
There is great confusion in the land - runtimeConfig
is used both in the spec.json and as a marged entity
(the thing actually used for running a db).

So:
runtimeConfigSpec is the settings in the specs.json
file. It's stored on disk.

runtimeConfig is the settings merged in memory
(using runtimeConfigSpec:s from disk and the
default settings from an engine).

So to repeat: runtimeConfigSpecs live on disk,
runtimeConfigs live in memory.

Test plan:
* Verify appendRuntimeConfig to a spec
  updates it's runtimeConfigSpec.
* sider snapshot add crates a
  specfile with an empty runtimeConfigSpec.
* sider snapshot add ... port=123
  persists the settings to the runtimeConfigSpec.
* sider db clone creates a file with empty
  runtimeConfigSpec property.
* sider db start port=123 uses the port
  but does not persist it.
* sider db start -p persists the settings
  to the propterty runtimeConfigSpec.
* sider engine getconf returns settings
  on a specfile having the runtimeConfigSpec.
* sider engine setconf saves properties on
  an object with a runtimeConfigSpec.