Migration: oops, re-do engines a bit
Engines weren't cleaning up any
non-standard folders. Re-do it a bit
so we do all in one fell swoop and
if the storage folder is non-standard, remove
everything in the end.

Also:
In v0_0_8 rc - Add ensure folder
so we can compare the standard-folders
as the strings dbs/ snapshots/ engines/

Test plan:
* Run migration on a new standard folder - verify
  config.json is removed and replaced with spec.json.
* Run migration on a non-standard folder and
  verify the whole non-standard folder is removed.