Add sider db list -s
Adding back -s (for displaying settings) for each
database.

In order for this to work, the table needs to handle
multilines, and also in order to have coloring
in the table and still be able to dynamically
shrink it, it needs to be ansi escape sequence
aware (so it can ellipsize and still keep coloring)
which is hard to do with the post-processing step.

Test plan:
* Have to databases. Verify that sider db list
  still works.
* Verify that sider db list -s works with no databases.
* Add some database. Verify that sider db list -s
  now displays a settings column. Make some
  new settings on the db and verify that the
  new settings are displayed in the settings column.
* Type sider db list <tab><tab>. Verify that -s
  and --settings are included in the completion.