Postgres: add password env variable or disable trust
Postgres now either requires a POSTGRES_PASSWORD enviroment
variable to be set when starting the docker image, or this
can be circumvented by seting POSTGRES_HOST_AUTH_METHOD=trust
for which passwords are ignored. Default to ignoring passwords
because this is for fast local development and not production, but
also add in the possiblity to set password with sider engine setconf.

Test plan:
* Have no password set in the engine settings (sider engine
  postgres getconf) lists no password=pwd. Run sider
  snapshot empty postgres goat and verify that snapshot
  starts. Verify you can connect to it with psql -U postgres
  -h localhost without a password.
* Set a password in the engine conf via sider engine postgres
  setconf password=pwd. Add an empty snapshot based on postgres.
  Verify the database can be connected to via PGPASSWORD=pwd
  psql -U postgres -h localhost.