Add mv (rename) for databases
It's been a long time coming, but
getting the database name right the first
time is hard. Add in mv / rename.

It checks for renaming to same name, if
the db exists, if the move to name
already exists.

If none of those apply, it renames it
by simply renaming the folder in which
the db resides.

Test plan:
* Verify that rename to same name gives an error.
* Verify that rename when the db does not exists
  gives an error.
* Verify that rename to an existing other db gives
  an error.
* Verify that renaming a valid db to a valid
  name changes the folder name in ~/.sider/dbs
  and that sider db list shows the new name.