Fix sider snapshot broken without specs file
Previously sider assumed that all the directories
inside the snapshots base folder were valid
snapshots and was designed around the fact
that no broken folders should be in there.

That was wrong. The most glaring example
is when running
`sider snapshot redis empty example`
and in another terminal issuing
`sider snapshot list`, the empty in-progress
snapshot breaks the listing.

It could also be cloned even though the
empty uninitialized database has not
yet shut down.

In order to fix this the design needs to
shift so that the specs.json file is the
marker when the snapshot is valid.

Ergo: check snapshots folders with a
specs.json file in it, if it's there
we're good to party like its 1989.

This means there can be clutter and
invalid folders in the snapshots/ folder.
But this is a trade-off worth making to
keep sider stable.

Test plan:
* Start an empty database (e g
  sider empty redis some-name) and
  leave it running in one terminal.
  Switch to another terminal and
  do sider snapshot list.
  Verify that the snapshot in progress
  does not show in the list.
* Try to clone out the snapshot in progress.
  Verify that there is an error with
  snapshot not found.
* Run completion with the snapshot
  in progress and verify it does not
  show up in the list.