Oh shit: don't save synthetic snapshot properties
But it did. sider snapshot mv renamed the folder (as should),
but not the erroneously saved synthetic property snapshotName
(derived from the folder-name for ease of use). This
cased completion on sider snapshot <cmd> to complete
correctly with the folder name as the snapshot name,
but sider snapshot list to display the saved
synthetic old snapshot name before the move.

There since v0.0.8 - oh it's a migration :(

Test plan:
* Delete the saved properties on a snapshot
  (they all have it since v0.0.8). Do a sider
  mv on a snapshot and verify that sider snapshot
  list now displays the new name instead of the
  name on which the snapshot was created with.