Npm 7+ doesn't print output in postinstall
The migrations built on the fact that you could
speak to the user when running the postinstall
hook. Not so since npm7.

https://github.com/npm/cli/issues/3347#issuecomment-854032079

Change strategy for the migrations. Don't ask.
Just do. If a migration encounters an error
any console.log, console.error plus the
error-message itself will be written to the
console + the npm install debug log.

Rely on this and change any process.exit(1) to
throw instead. We can then catch any error at
the top level, print some nice message and
exit out cleanly.

First up is the db.js in the v0.0.8 migration.
Re-word so the base error is consistently included,
remove any process.exit(1).

Don't try to half-do migrations, either they
all apply cleanly or we bail.

Test plan:
* node migrations/v0.0.8-v1.0.0/dbs.js && echo $?
  is zero.